Moved blogs into a separate file
