+++ app/data/blogPosts.ts
@@ -1,0 +1,74 @@
+const blogPosts = [
+  {
+    id: 1,
+    image: "/blogs/blog1_placeholder.jpg",
+    date: "April 4, 2025",
+    location: "Toronto, Canada",
+    title: "Mastering the Basics of Karate",
+    description:
+      "Discover how mastering the fundamentals can lead to better results in both training and competitions.",
+    slug: "mastering-the-basics-of-karate",
+    content: `
+      This is the full content of the blog post. 
+      You can write HTML/JSX-safe strings here, or convert from markdown later if needed.
+    `,
+  },
+  {
+    id: 2,
+    image: "/blogs/blog2_placeholder.jpg",
+    date: "March 25, 2025",
+    location: "Vancouver, Canada",
+    title: "The Belt Journey Explained",
+    description:
+      "Learn about the meaning behind each karate belt and what it takes to move up the ranks.",
+    slug: "belt-journey-explained",
+    content: `
+    This is the full content of the blog post. 
+    You can write HTML/JSX-safe strings here, or convert from markdown later if needed.
+  `,
+  },
+  {
+    id: 3,
+    image: "/blogs/blog3_placeholder.jpg",
+    date: "March 10, 2025",
+    location: "Calgary, Canada",
+    title: "Training Techniques for Beginners",
+    description:
+      "Start your karate journey with these essential techniques and mindset tips.",
+    slug: "training-techniques-for-beginners",
+    content: `
+    This is the full content of the blog post. 
+    You can write HTML/JSX-safe strings here, or convert from markdown later if needed.
+  `,
+  },
+  {
+    id: 4,
+    image: "/blogs/blog2_placeholder.jpg",
+    date: "February 28, 2025",
+    location: "Montreal, Canada",
+    title: "Karate Sparring Tips",
+    description:
+      "Enhance your sparring skills with these expert tips and strategies.",
+    slug: "karate-sparring-tips",
+    content: `
+    This is the full content of the blog post. 
+    You can write HTML/JSX-safe strings here, or convert from markdown later if needed.
+  `,
+  },
+  {
+    id: 5,
+    image: "/blogs/blog1_placeholder.jpg",
+    date: "February 15, 2025",
+    location: "Ottawa, Canada",
+    title: "The Importance of Kata in Karate",
+    description:
+      "Understand the significance of kata in karate training and its role in competitions.",
+    slug: "importance-of-kata-in-karate",
+    content: `
+    This is the full content of the blog post. 
+    You can write HTML/JSX-safe strings here, or convert from markdown later if needed.
+  `,
+  },
+];
+
+export default blogPosts;
