+++ app/sections/BlogGallery.tsx
@@ -1,59 +1,7 @@
 import Image from "next/image";
 import Link from "next/link";
 import { motion } from "framer-motion";
-
-const blogPosts = [
-  {
-    id: 1,
-    image: "/blogs/blog1_placeholder.jpg",
-    date: "April 4, 2025",
-    location: "Toronto, Canada",
-    title: "Mastering the Basics of Karate",
-    description:
-      "Discover how mastering the fundamentals can lead to better results in both training and competitions.",
-    slug: "mastering-the-basics-of-karate",
-  },
-  {
-    id: 2,
-    image: "/blogs/blog2_placeholder.jpg",
-    date: "March 25, 2025",
-    location: "Vancouver, Canada",
-    title: "The Belt Journey Explained",
-    description:
-      "Learn about the meaning behind each karate belt and what it takes to move up the ranks.",
-    slug: "belt-journey-explained",
-  },
-  {
-    id: 3,
-    image: "/blogs/blog3_placeholder.jpg",
-    date: "March 10, 2025",
-    location: "Calgary, Canada",
-    title: "Training Techniques for Beginners",
-    description:
-      "Start your karate journey with these essential techniques and mindset tips.",
-    slug: "training-techniques-for-beginners",
-  },
-  {
-    id: 4,
-    image: "/blogs/blog2_placeholder.jpg",
-    date: "February 28, 2025",
-    location: "Montreal, Canada",
-    title: "Karate Sparring Tips",
-    description:
-      "Enhance your sparring skills with these expert tips and strategies.",
-    slug: "karate-sparring-tips",
-  },
-  {
-    id: 5,
-    image: "/blogs/blog1_placeholder.jpg",
-    date: "February 15, 2025",
-    location: "Ottawa, Canada",
-    title: "The Importance of Kata in Karate",
-    description:
-      "Understand the significance of kata in karate training and its role in competitions.",
-    slug: "importance-of-kata-in-karate",
-  },
-];
+import blogPosts from "../data/blogPosts";
 
 export default function BlogGallery() {
   return (
